validated ploss. Removed warnings and cleaned code. Added msq_lossy

+++ pmcsn/src/analytic.h
@@ -23,7 +23,7 @@ double get_theoretical_visits(block_type type, int num_servers);
 double  erlang_c_block_probability(int m, double rho);
 double  erlang_c_queue_time(double block_probability, double service_time_multi, double rho);
 double  erlang_c_response_time(double queue_time, double service_time);
-double  erlang_b_loss_probability(int m, double lambda, double mhu);
+long double erlang_b_loss_probability(int m, double lambda, double mhu);
 // void calculate_interval_estimate_for_stat(stat_type stat, const char *stat_name, replica_stats *replica_stats_ensemble, const char *block_name);
 void calculate_interval_estimate_for_stat(const char *stat_name, const double *global_response_time, int num_replicas);
 double get_theoretical_response_time(block_type type, int m);
